Moved navigation test to App.test
