+++ src/components/Home.test.js
@@ -1,10 +1,9 @@
 import React from "react";
-import { render, screen } from "@testing-library/react";
+import { render } from "@testing-library/react";
 import "@testing-library/jest-dom/extend-expect";
-import userEvent from "@testing-library/user-event";
-import Home from "./Home.js";
-import App from "../App.js";
 import { BrowserRouter as Router } from "react-router-dom";
+
+import Home from "./Home.js";
 
 test("loads the home page", async () => {
   const { getByText } = render(
@@ -14,13 +13,4 @@ test("loads the home page", async () => {
   );
   const titleElement = getByText(/TK Recipe App/i);
   expect(titleElement).toBeInTheDocument();
-});
-
-test("access recipies", async () => {
-  render(<App />);
-  const leftClick = { button: 0 };
-  userEvent.click(screen.getByRole("link"), leftClick);
-  expect(
-    await screen.findByText(/This is the list of recipes/i)
-  ).toBeInTheDocument();
 });
